simplified CarveHole
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@797 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/CMap.h
@@ -218,7 +218,7 @@ public:
     void        PutImagePixel(uint x, uint y, Uint32 colour);
 
 	void		UpdateMiniMap(bool force = false);
-	void		UpdateMiniMapRect(ushort x, ushort y, ushort w, ushort h);	
+	void		UpdateMiniMapRect(int x, int y, int w, int h);	
 
 	void		Send(CBytestream *bs);
